Infra/v9 tests (#3903)
* Fix button and update @testing-library/react-native version

* Remove unsupported modifiers methods, add docs and fix tests

The old TextField was the only usage of the removed methods and it was only used by MaskedInput which is deprecated
