+++ docs/getting-started/v8.md
@@ -1,5 +1,5 @@
 ---
-sidebar_position: 5
+sidebar_position: 6
 sidebar_label: Migrating v7 -> v8
 title: "Migrating v7 -> v8"
 # path: "/getting-started/v8"
@@ -7,6 +7,7 @@ title: "Migrating v7 -> v8"
 ## `react-native-ui-lib@8.x.x`
 
 ## General
+Now supports react-native 0.77  
 `uilib-native` (our native library) has been moved from `dependencies` to `peerDependencies`.  
 Make sure to `pod install` after updating.  
 We do plan on making this optional in the future.
