+++ docs/getting-started/v9.md
@@ -1,0 +1,21 @@
+---
+sidebar_position: 7
+sidebar_label: Migrating v8 -> v9
+title: "Migrating v8 -> v9"
+# path: "/getting-started/v9"
+---
+## `react-native-ui-lib@9.x.x`
+
+## General
+Now supports react-native 0.78 and React 19  
+
+## Components
+
+### MaskedInput
+Only the newer version is now available (the `migrate` prop is removed)  
+
+## Utils
+
+### modifiers
+extractOwnProps - removed  
+extractComponentProps - removed
