+++ package.json
@@ -10,7 +10,8 @@
     "ios": "yarn workspace react-native-ui-lib ios",
     "android": "yarn workspace react-native-ui-lib android",
     "iPad": "yarn workspace react-native-ui-lib iPad",
-    "test": "yarn lint && yarn workspace react-native-ui-lib test",
+    "test": "yarn workspace react-native-ui-lib test",
+    "pretest": "yarn lint",
     "lint": "eslint packages -c .eslintrc.js --ext .tsx,.ts,.js",
     "lint:fix": "eslint packages -c .eslintrc.js --fix",
     "build:dev": "tsc --p tsconfig.dev.json",
@@ -40,7 +41,7 @@
     "@react-native/metro-config": "0.78.3",
     "@react-native/typescript-config": "0.78.3",
     "@shopify/flash-list": "1.7.6",
-    "@testing-library/react-native": "^11.5.1",
+    "@testing-library/react-native": "^13.3.3",
     "@types/hoist-non-react-statics": "^3.3.7",
     "@types/jest": "^29.5.13",
     "@types/lodash": "^4.0.0",
