+++ packages/react-native-ui-lib/package.json
@@ -64,7 +64,7 @@
         "@react-native/metro-config": "0.78.3",
         "@react-native/typescript-config": "0.78.3",
         "@shopify/flash-list": "1.7.6",
-        "@testing-library/react-native": "^11.5.1",
+        "@testing-library/react-native": "^13.3.3",
         "@types/hoist-non-react-statics": "^3.3.7",
         "@types/jest": "^29.5.13",
         "@types/lodash": "^4.0.0",
