+++ packages/react-native-ui-lib/src/components/maskedInput/maskedInput.api.json
@@ -44,7 +44,6 @@
     "  return (",
     "    <div>",
     "      <MaskedInput",
-    "        migrate",
     "        renderMaskedText={renderTimeText}",
     "        formatter={formatter}",
     "        keyboardType={'numeric'}",
