+++ packages/react-native-ui-lib/src/components/picker/types.ts
@@ -192,7 +192,6 @@ export type PickerBaseProps = Omit<TextFieldProps, 'value' | 'onChange'> &
   PickerPropsDeprecation &
   PickerExpandableOverlayProps &
   PickerListProps & {
-    /* ...TextField.propTypes, */
     /**
      * Use dialog instead of modal picker
      */
